# apps/api-gateway/openapi.yaml
openapi: 3.1.0
info:
  title: AGRORED API Gateway
  version: 0.1.0
  description: Punto de entrada del ecosistema AGRORED para enrutar peticiones a los microservicios.
servers:
  - url: http://localhost:8080
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
security:
  - BearerAuth: []
paths:
...
  /api/v1/inventory/register:
    post:
      summary: Registrar item de inventario
      tags: [Inventory]
      responses:
        "201":
          description: Inventario registrado
        "400":
          description: Payload invalido
  /api/v1/inventory:
    get:
      summary: Listar inventario
      tags: [Inventory]
      parameters:
        - name: page
          in: query
          schema: { type: integer, default: 1 }
        - name: limit
          in: query
          schema: { type: integer, default: 20, maximum: 100 }
      responses:
        "200":
          description: Lista paginada de inventario
  /api/v1/inventory/{id}:
    get:
      summary: Obtener item de inventario por ID
      tags: [Inventory]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: Detalle del inventario
        "404":
          description: No encontrado
  /api/v1/logistics/register:
    post:
      summary: Registrar orden logistica
      tags: [Logistics]
      responses:
        "201":
          description: Orden registrada
        "400":
          description: Payload invalido
  /api/v1/logistics:
    get:
      summary: Listar ordenes logisticas
      tags: [Logistics]
      parameters:
        - name: page
          in: query
          schema: { type: integer, default: 1 }
        - name: limit
          in: query
          schema: { type: integer, default: 20, maximum: 100 }
      responses:
        "200":
          description: Lista paginada de ordenes
  /api/v1/logistics/{id}:
    get:
      summary: Obtener orden logistica por ID
      tags: [Logistics]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: Detalle de la orden
        "404":
          description: No encontrado
...